+++ CHANGELOG.md
@@ -12,6 +12,13 @@ a [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Se incrementa **MINOR**, cuando se agregue una o más **nueva(s) funcionalidad(es) a la aplicación**.
 * Se incrementa **PATCH**, cuando se implemente un **fix**.
 
+## [0.0.3] - 2026-04-13
+### [TR]
+* FIX - Plantilla para carga de trabajadores.
+* FIX - Ajustar lógica de modal error y éxito en facturas masivas.
+* FIX - Error al actualizar/editar UsoCFDI en importar facturas masivas.
+* FIX - Error al editar de facturas información global.
+
 ## [0.0.2] - 2026-03-26
 ### [TR]
 * FIX - Ajuste en sección de nóminas y validaciones de conceptos